+++ internal/bytesutil/bytesutil.go
@@ -64,9 +64,11 @@ func PutBytes(b *[]byte) {
 	bytePool.Put(b)
 }
 
-// ParseInt is a specialized version of strconv.ParseInt that parses a base-10 encoded signed integer from a []byte.
+// ParseInt is a specialized version of strconv.ParseInt that parses a base-10
+// encoded signed integer from a []byte.
 //
-// This can be used to avoid allocating a string, since strconv.ParseInt only takes a string.
+// This can be used to avoid allocating a string, since strconv.ParseInt only
+// takes a string.
 func ParseInt(b []byte) (int64, error) {
 	if len(b) == 0 {
 		return 0, errors.New("empty slice given to parseInt")
@@ -90,9 +92,11 @@ func ParseInt(b []byte) (int64, error) {
 	return int64(n), nil
 }
 
-// ParseUint is a specialized version of strconv.ParseUint that parses a base-10 encoded integer from a []byte.
+// ParseUint is a specialized version of strconv.ParseUint that parses a base-10
+// encoded integer from a []byte.
 //
-// This can be used to avoid allocating a string, since strconv.ParseUint only takes a string.
+// This can be used to avoid allocating a string, since strconv.ParseUint only
+// takes a string.
 func ParseUint(b []byte) (uint64, error) {
 	if len(b) == 0 {
 		return 0, errors.New("empty slice given to parseUint")
@@ -112,20 +116,30 @@ func ParseUint(b []byte) (uint64, error) {
 	return n, nil
 }
 
-// Expand expands the given byte slice to exactly n bytes.
-//
-// If cap(b) < n, a new slice will be allocated and filled with the bytes from b.
-func Expand(b []byte, n int) []byte {
-	if cap(b) < n {
+func expand(b []byte, n int, keepBytes bool) []byte {
+	if n == 0 && b == nil {
+		b = []byte{} // so as to not return nil
+	} else if cap(b) < n {
 		nb := make([]byte, n)
-		copy(nb, b)
+		if keepBytes {
+			copy(nb, b)
+		}
 		return nb
 	}
 	return b[:n]
 }
 
-// BufferedBytesDelim reads a line from br and checks that the line ends with \r\n, returning the line without \r\n.
-func BufferedBytesDelim(br *bufio.Reader) ([]byte, error) {
+// Expand expands the given byte slice to exactly n bytes. It will not return
+// nil.
+//
+// If cap(b) < n then a new slice will be allocated.
+func Expand(b []byte, n int) []byte {
+	return expand(b, n, false)
+}
+
+// ReadBytesDelim reads a line from br and checks that the line ends with
+// \r\n, returning the line without \r\n.
+func ReadBytesDelim(br *bufio.Reader) ([]byte, error) {
 	b, err := br.ReadSlice('\n')
 	if err != nil {
 		return nil, err
@@ -135,9 +149,10 @@ func BufferedBytesDelim(br *bufio.Reader) ([]byte, error) {
 	return b[:len(b)-2], err
 }
 
-// BufferedIntDelim reads the current line from br as an integer.
-func BufferedIntDelim(br *bufio.Reader) (int64, error) {
-	b, err := BufferedBytesDelim(br)
+// ReadIntDelim reads the current line from br as an integer, checks that the
+// line ends with \r\n, and returns the integer.
+func ReadIntDelim(br *bufio.Reader) (int64, error) {
+	b, err := ReadBytesDelim(br)
 	if err != nil {
 		return 0, err
 	}
@@ -150,7 +165,7 @@ func ReadNAppend(r io.Reader, b []byte, n int) ([]byte, error) {
 		return b, nil
 	}
 	m := len(b)
-	b = Expand(b, len(b)+n)
+	b = expand(b, len(b)+n, true)
 	_, err := io.ReadFull(r, b[m:])
 	return b, err
 }